Add file handling for diary entries and exit option
Included `minDagbok.txt` in the project with `PreserveNewest` to ensure it is copied to the output directory.

Added `SparaDagbokFil` method in `DagbokHanterare` to save diary entries to a file in "Date|Text" format.

Replaced placeholder in case "4" of `Program.cs` with a call to `LäsDagbokFil` to enable reading diary entries from the file.

Added case "5" to allow the user to exit the program with a farewell message.